Fix(bumsoo/promotion): 대댓글 겹쳐서 달리는 오류 해결

+++ src/apis/promotion.js
@@ -35,46 +35,41 @@ export function deletePost(postId) {
   return axios.delete(`api/board/promotion/${postId}`);
 }
 
-export function editComment(postId, commentNo, description) {
-  return axios.put(`api/board/promotion/${postId}/${commentNo}`, {
+export function editComment(commentNo, parentCommentID, description, postId) {
+  if (parentCommentID) {
+    return axios.put(
+      `/api/board/promotion/${postId}/${parentCommentID}/${commentNo}`,
+      { description }
+    );
+  }
+  return axios.put(`/api/board/promotion/${postId}/${commentNo}`, {
     description
   });
 }
 
-export function editReplyComment(
-  postId,
-  commentNo,
-  replyCommentNo,
-  description
-) {
-  return axios.put(
-    `api/board/promotion/${postId}/${commentNo}/${replyCommentNo}`,
-    {
-      description
-    }
-  );
-}
-
-export function deleteComment(postId, commentNo) {
+export function deleteComment(commentNo, parentCommentID, postId) {
+  console.log(postId);
+  if (parentCommentID) {
+    return axios.delete(
+      `/api/board/promotion/${postId}/${parentCommentID}/${commentNo}`
+    );
+  }
   return axios.delete(`api/board/promotion/${postId}/${commentNo}`);
 }
 
-export function deleteReplyComment(postId, commentNo, replyCommentNo) {
-  return axios.delete(
-    `api/board/promotion/${postId}/${commentNo}/${replyCommentNo}`
-  );
-}
-
-export function addComment(postId, description) {
+export function addComment(postId, description, parentCommentID) {
+  if (parentCommentID) {
+    return axios.post(`/api/board/promotion/${postId}/${parentCommentID}`, {
+      id: 'test1',
+      description,
+      url: `/promotion`,
+      notiCategoryNum: 1
+    });
+  }
   return axios.post(`api/board/promotion/${postId}`, {
     id: 'test1',
-    description
-  });
-}
-
-export function replyAddComment(postId, parentCommentId, description) {
-  return axios.post(`api/board/promotion/${postId}/${parentCommentId}`, {
-    id: 'test1',
-    description
+    description,
+    url: `/promotion`,
+    notiCategoryNum: 0
   });
 }
